style final restructure
completed styling restructure to more modular model
