+++ github-project/src/pages/Profile/Profile.module.css
@@ -1,0 +1,165 @@
+
+.container {
+    min-height: 100vh;
+    height: 100%;
+  }
+  
+  .profileContainer {
+    display: grid;
+    grid-template-columns: 1fr 2fr;
+    grid-template-rows: auto auto 1fr;
+    gap: 1rem;
+    flex-grow: 1;
+    min-height: 100vh;
+  }
+  
+  :global(body.dark-mode) {
+        .profileContainer {
+            background-color: var(--green-dark);
+        }
+    }
+  
+  .profileTitle {
+    grid-column: 1 / 3;
+  }
+  
+  .profileInfo {
+    margin-top: 5rem;
+    grid-column: 1;
+    grid-row: 2;
+    justify-self: center;
+  }
+  
+  .statisticDropdown {
+    grid-column: 1;
+    grid-row: 3;
+    justify-self: center;
+  }
+  
+  .selectedStatistic {
+    padding-top: 5vh;
+    display: flex;
+    flex-direction: column; 
+    align-items: stretch;   
+    text-align: center;     
+    gap: 2rem;
+  }
+  
+  .yearSelector {
+    padding-bottom: 3rem;
+  }
+  
+  .chartWrapper {
+    width: 90%;
+  }
+  
+  .statisticDropdown select {
+    background-color: var(--white);
+    border: 2px solid var(--green);
+    padding: 0.5rem;
+    color: var(--green);
+    width: clamp(200px, 10vw, 250px);
+  }
+  
+  :global(body.dark-mode) {
+    .statisticDropdown select {
+        background-color: var(--green-dark);
+        color: var(--white);
+        border-color: var(--white-50);
+      }
+  } 
+  
+  .selectedStatistic {
+    grid-column: 2;
+    grid-row: 2 / 4;
+    overflow: hidden;
+  }
+  
+  :global(body.dark-mode) {
+    .contributionContainer {
+        background-color: var(--green-dark);
+      }
+  } 
+  
+  .chartContainer {
+    width: 100%;
+    height: auto;
+    overflow: visible;
+    padding: 10px;
+    box-sizing: border-box;
+    border-radius: 10px;
+    background-color: #1e1e1e;
+    box-shadow: 0 0 10px rgba(255, 255, 255, 0.2);
+  }
+  
+  :global(body.dark-mode) {
+    .chartContainer {
+        background-color: var(--green-dark);
+      }
+  }
+  
+  .chartContainer svg {
+    stroke: white !important;
+    color: white !important;
+  }
+  
+  .chartContainer text {
+    fill: white !important;
+  }
+  
+  .languagePieChart {
+    margin: auto auto;
+    text-align: center;
+    width: 100%;
+    height: auto;
+    max-width: 600px;
+    padding: 1vh 20px;
+    overflow: auto;
+    justify-items: center;
+  }
+  
+  .pieContainer {
+    display: flex;
+    flex-direction: row;
+  }
+  
+  .recharts-cartesian-axis-line,
+  .recharts-cartesian-tick-line,
+  .recharts-cartesian-grid line {
+    stroke: var(--green);
+  }
+  
+  .recharts-cartesian-axis-tick text {
+    fill: var(--green);
+  }
+  
+  :global(body.dark-mode) {
+    .recharts-cartesian-axis-line,
+    .recharts-cartesian-tick-line,
+    .recharts-cartesian-grid line {
+        stroke: var(--white);
+    }
+  }
+
+  :global(body.dark-mode) {
+    .recharts-cartesian-axis-tick text {
+        fill: var(--white);
+      }
+  } 
+  
+  .recharts-legend-wrapper {
+    padding: 2rem;
+    background-color: var(--lime);
+    border-radius: 20px;
+  }
+  
+  .recharts-default-legend {
+    display: flex;
+    flex-direction: row;
+  }
+  
+  .recharts-legend-item-text {
+    color: var(--green) !important;
+    font-weight: 900;
+  }
+
