+++ github-project/src/styles/global.css
@@ -175,161 +175,7 @@ main {
   border-radius: 10px; 
 }
 
-/* User Profile Styling */
-
-.container {
-  min-height: 100vh;
-  height: 100%;
-}
-
-.profile-container {
-  display: grid;
-  grid-template-columns: 1fr 2fr;
-  grid-template-rows: auto auto 1fr;
-  gap: 1rem;
-  flex-grow: 1;
-  min-height: 100vh;
-}
-
-body.dark-mode .profile-container {
-  background-color: var(--green-dark);
-}
-
-.profile-title {
-  grid-column: 1 / 3;
-}
-
-.profile-info {
-  margin-top: 5rem;
-  grid-column: 1;
-  grid-row: 2;
-  justify-self: center;
-}
-
-.statistic-dropdown {
-  grid-column: 1;
-  grid-row: 3;
-  justify-self: center;
-}
-
-.selected-statistic {
-  padding-top: 5vh;
-  display: flex;
-  flex-direction: column; 
-  align-items: stretch;   
-  text-align: center;     
-  gap: 2rem;
-}
-
-.year-selector {
-  padding-bottom: 3rem;
-}
-
-.chart-wrapper {
-  width: 90%;
-}
-
-.statistic-dropdown select {
-  background-color: var(--white);
-  border: 2px solid var(--green);
-  padding: 0.5rem;
-  color: var(--green);
-  width: clamp(200px, 10vw, 250px);
-}
-
-body.dark-mode .statistic-dropdown select {
-  background-color: var(--green-dark);
-  color: var(--white);
-  border-color: var(--white-50);
-}
-
-.selected-statistic {
-  grid-column: 2;
-  grid-row: 2 / 4;
-  overflow: hidden;
-}
-
-body.dark-mode .contribution-container {
-  background-color: var(--green-dark);
-}
-
-.chart-container {
-  width: 100%;
-  height: auto;
-  overflow: visible;
-  padding: 10px;
-  box-sizing: border-box;
-  border-radius: 10px;
-  background-color: #1e1e1e;
-  box-shadow: 0 0 10px rgba(255, 255, 255, 0.2);
-}
-
-body.dark-mode .chart-container {
-  background-color: var(--green-dark);
-}
-
-.chart-container svg {
-  stroke: white !important;
-  color: white !important;
-}
-
-.chart-container text {
-  fill: white !important;
-}
-
-.language-pie-chart {
-  margin: auto auto;
-  text-align: center;
-  width: 100%;
-  height: auto;
-  max-width: 600px;
-  padding: 1vh 20px;
-  overflow: auto;
-  justify-items: center;
-}
-
-.pie-container {
-  display: flex;
-  flex-direction: row;
-}
-
-.recharts-cartesian-axis-line,
-.recharts-cartesian-tick-line,
-.recharts-cartesian-grid line {
-  stroke: var(--green);
-}
-
-.recharts-cartesian-axis-tick text {
-  fill: var(--green);
-}
-
-body.dark-mode .recharts-cartesian-axis-line,
-body.dark-mode .recharts-cartesian-tick-line,
-body.dark-mode .recharts-cartesian-grid line {
-  stroke: var(--white);
-}
-
-body.dark-mode .recharts-cartesian-axis-tick text {
-  fill: var(--white);
-}
-
-.recharts-legend-wrapper {
-  padding: 2rem;
-  background-color: var(--lime);
-  border-radius: 20px;
-}
-
-.recharts-default-legend {
-  display: flex;
-  flex-direction: row;
-}
-
-.recharts-legend-item-text {
-  color: var(--green) !important;
-  font-weight: 900;
-}
-
 .error {
-  color: red;
+
   margin-top: 10px;
 }
